Refactor code into Ui, Storage, Parser, TaskList
