+++ src/main/data/SavedList.txt
@@ -1,6 +1,5 @@
 T|1|yes
 T|0|yeayea
-D|0|yes|2024-12-15
-D|0|yes|2024-12-15
-D|0|yes|2024-11-15
-E|0|yes|2024-11-16|2025-11-16
+T|0|this
+D|0|test1|2019-03-29
+E|0|sleep|2024-11-14|2024-11-20
