+++ src/main/java/Commands.java
@@ -1,4 +1,4 @@
-enum Commands {
+public enum Commands {
     HELP,
     TODO,
     DEADLINE,
