+++ src/main/java/Ui.java
@@ -1,0 +1,17 @@
+public class Ui {
+
+    private static String dividerText = "____________________________________________________________\n";
+
+    public static void greet() {
+        String greetText = "Hello! I'm Tam the Task Manager!\nWhat can I do for you?\n";
+        System.out.print(dividerText);
+        System.out.print(greetText);
+        System.out.print(dividerText);
+    }
+
+    public static void exit() {
+        String exitText = "Bye. Hope to see you again soon!\n";
+        System.out.print(exitText);
+        System.out.print(dividerText);
+    }
+}
